requirements/base.txt
Django==4.1.2
django-auth-ldap==4.1.0
django-bootstrap-modal-forms==2.2.0
django-crum==0.7.9
django-filter==22.1
django-pwa==1.0.10
django-widget-tweaks==1.4.12